src/test/java/com/v/SimulationTest.java
package com.v;

import org.junit.jupiter.api.Test;
import static org.junit.jupiter.api.Assertions.*;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.AfterEach;
import static org.mockito.Mockito.*;

public class SimulationTest {
    private Simulation simulation;
    private int[][] adjacencyMatrix = {
        {0, 1, 0},
        {1, 0, 1},
        {0, 1, 0}
    };

    @BeforeEach
    void setUp() {
        simulation = new Simulation();
    }

    @AfterEach
    void tearDown() {
        // Clean up resources if needed
    }

    @Test
    void testSimulationRuns() {
        String result = simulation.runSimulation(3, adjacencyMatrix);
        assertNotNull(result);
        assertTrue(result.contains("Simulation"));
    }

    @Test
    void testSimulationWithNoNodes() {
        // Assuming Simulation can accept node count as a parameter
        String result = simulation.runSimulation(0, adjacencyMatrix);
        assertNotNull(result);
        assertTrue(result.contains("No nodes to simulate"));
    }

    @Test
    void testSimulationWithSingleNode() {
        String result = simulation.runSimulation(1, adjacencyMatrix);
        assertNotNull(result);
        assertTrue(result.contains("Simulation completed with 1 node"));
    }

    @Test
    void testSimulationWithMultipleNodes() {
        String result = simulation.runSimulation(5, adjacencyMatrix);
        assertNotNull(result);
        assertTrue(result.contains("Simulation completed with 5 nodes"));
    }

    @Test
    void testNodeConnections() {
        Simulation simulation = spy(new Simulation());
        simulation.runSimulation(3, adjacencyMatrix);
        verify(simulation, atLeastOnce()).runSimulation(3, adjacencyMatrix);
        // Additional verifications can be added here
    }
}
